+++ changelog.md
@@ -5,6 +5,9 @@ __                                                                              
 * New Map
 * Decals
 * River Push Loot
+* Vehicles Rework
+* Boat
+* Vehicles Tire Marks
 __                                                                                                                 __
 ### Menu
 __                                                                                                                 __